requirements.txt
pytest
pytest-cov
pytest-mock
black
isort
flake8
mypy
pre-commit
discord.py
tinydb
configparser